third_party/blink/web_tests/virtual/pna-workers-disabled/external/wpt/fetch/private-network-access/service-worker-fetch.tentative.https.window_9-last-expected.txt

This is a testharness.js-based test.
[FAIL] treat-as-public to local: failed preflight.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[FAIL] treat-as-public to local: success.
  assert_equals: fetch error expected (undefined) undefined but got (string) "TypeError"
[FAIL] treat-as-public to private: failed preflight.
  assert_equals: fetch error expected (string) "TypeError" but got (undefined) undefined
[FAIL] treat-as-public to private: success.
  assert_equals: fetch error expected (undefined) undefined but got (string) "TypeError"
Harness: the test ran to completion.
